requirements.txt
Django==2.1.7
sympy==1.3
scipy==1.2.1
opencv-python==4.0.0.21
Shapely==1.6.4.post2
face-recognition==1.2.3
matplotlib==3.0.3
django-picklefield==2.0